moved complex opcodes
