+++ Engine/entry.c
@@ -126,7 +126,7 @@ OENTRY opcodlst_1[] = {
   { "=.r",    S(ASSIGN),0,        "r",    "i",    rassign, NULL, NULL, NULL },
   { "=.Ci",    S(R2CXOP),0,        ":Complex;",    "iio", (SUBR) complex_assign},
   { "=.C",    S(CXOP),0,        ":Complex;",    ":Complex;", (SUBR) complex_init,
-                                                                (SUBR) complex_init},
+    (SUBR) complex_init},
   { "=.Ck",    S(R2CXOP),0,        ":Complex;",    "kkO", NULL, (SUBR) complex_assign},
   { "=.i",    S(ASSIGNM),0,       "IIIIIIIIIIIIIIIIIIIIIIII", "m",
     minit, NULL, NULL, NULL  },
@@ -156,7 +156,7 @@ OENTRY opcodlst_1[] = {
   { "delete", S(AOP),0, "", ":Opcode;[]", NULL, NULL, (SUBR) opcode_delete_array},
   { "opcodeinfo", S(OPINFO) ,0,  "", ":Opcode;", (SUBR) opcode_object_info},
   { "=.Opcode", S(ASSIGN) ,0,  ":Opcode;", ":Opcode;", (SUBR) copy_opcode_obj,
-  (SUBR) copy_opcode_obj},
+    (SUBR) copy_opcode_obj},
   { "run", S(OPRUN), 0, "*", ":Opcode;*", (SUBR) opcode_object_init, (SUBR) opcode_run_perf},
   { "run", S(OPRUN), 0, "*", ":Opcode;[]*", (SUBR) opcode_array_init, (SUBR) opcode_array_perf},
   { "init", S(OPRUN), 0, "*", ":Opcode;*", (SUBR) opcode_object_init},
@@ -202,50 +202,47 @@ OENTRY opcodlst_1[] = {
   { "##mul.ii",  S(AOP),0,          "i",    "ii",   mulkk                   },
   { "##div.ii",  S(AOP),0,          "i",    "ii",   divkk                   },
   { "##add.CC",  S(CXOP),0,          ":Complex;",    ":Complex;:Complex;",
-                                      (SUBR) complex_add, (SUBR) complex_add },
+    (SUBR) complex_add, (SUBR) complex_add },
   { "##sub.CC",  S(CXOP),0,          ":Complex;",    ":Complex;:Complex;",
-                                      (SUBR) complex_sub, (SUBR) complex_sub },
+    (SUBR) complex_sub, (SUBR) complex_sub },
   { "##mul.CC",  S(CXOP),0,          ":Complex;",    ":Complex;:Complex;",
-                                      (SUBR) complex_prod, (SUBR) complex_prod },
+    (SUBR) complex_prod, (SUBR) complex_prod },
   { "##div.CC",  S(CXOP),0,          ":Complex;",    ":Complex;:Complex;",
-                                      (SUBR) complex_div, (SUBR) complex_div },
+    (SUBR) complex_div, (SUBR) complex_div },
 
   { "##add.Ci",  S(AOP),0,          ":Complex;",    ":Complex;i",
-                                      (SUBR) complex_add_real},
+    (SUBR) complex_add_real},
   { "##add.Ck",  S(AOP),0,          ":Complex;",    ":Complex;k", NULL,
-                                      (SUBR) complex_add_real},
- { "##add.iC",  S(AOP),0,          ":Complex;",    "i:Complex;",
-                                      (SUBR) real_add_complex},
+    (SUBR) complex_add_real},
+  { "##add.iC",  S(AOP),0,          ":Complex;",    "i:Complex;",
+    (SUBR) real_add_complex},
   { "##add.kC",  S(AOP),0,          ":Complex;",    "k:Complex;", NULL,
-                                      (SUBR) real_add_complex},
+    (SUBR) real_add_complex},
 
   { "##sub.Ci",  S(AOP),0,          ":Complex;",    ":Complex;i",
-                                      (SUBR) complex_sub_real},
+    (SUBR) complex_sub_real},
   { "##sub.Ck",  S(AOP),0,          ":Complex;",    ":Complex;k", NULL,
-                                      (SUBR) complex_sub_real},
+    (SUBR) complex_sub_real},
   { "##sub.iC",  S(AOP),0,          ":Complex;",    "i:Complex;",
-                                      (SUBR) real_sub_complex},
+    (SUBR) real_sub_complex},
   { "##sub.kC",  S(AOP),0,          ":Complex;",    "k:Complex;", NULL,
-                                      (SUBR) real_sub_complex},
-
-  
+    (SUBR) real_sub_complex},
   { "##mul.Ci",  S(AOP),0,          ":Complex;",    ":Complex;i",
-                                      (SUBR) complex_mul_real},
+    (SUBR) complex_mul_real},
   { "##mul.Ck",  S(AOP),0,          ":Complex;",    ":Complex;k", NULL,
-                                      (SUBR) complex_mul_real},
- { "##mul.iC",  S(AOP),0,          ":Complex;",    "i:Complex;",
-                                      (SUBR) real_mul_complex},
+    (SUBR) complex_mul_real},
+  { "##mul.iC",  S(AOP),0,          ":Complex;",    "i:Complex;",
+    (SUBR) real_mul_complex},
   { "##mul.kC",  S(AOP),0,          ":Complex;",    "k:Complex;", NULL,
-                                      (SUBR) real_mul_complex},
-
+    (SUBR) real_mul_complex},
   { "##div.Ci",  S(AOP),0,          ":Complex;",    ":Complex;i",
-                                      (SUBR) complex_div_real},
+    (SUBR) complex_div_real},
   { "##div.Ck",  S(AOP),0,          ":Complex;",    ":Complex;k", NULL,
-                                      (SUBR) complex_div_real},
+    (SUBR) complex_div_real},
   { "##div.iC",  S(AOP),0,          ":Complex;",    "i:Complex;",
-                                      (SUBR) real_div_complex},
+    (SUBR) real_div_complex},
   { "##div.kC",  S(AOP),0,          ":Complex;",    "k:Complex;", NULL,
-                                      (SUBR) real_div_complex},
+    (SUBR) real_div_complex},
   { "##mod.ii",  S(AOP),0,          "i",    "ii",   modkk                   },
   { "##add.kk",  S(AOP),0,           "k",    "kk",   NULL,   addkk           },
   { "##sub.kk",  S(AOP),0,           "k",    "kk",   NULL,   subkk           },
@@ -275,6 +272,47 @@ OENTRY opcodlst_1[] = {
   { "##subin.k", S(ASSIGN),0,        "k",    "k",    NULL,   subin   },
   { "##subin.K", S(ASSIGN),0,        "a",    "k",    NULL,   subinak },
   { "##subin.a", S(ASSIGN),0,        "a",    "a",    NULL,   subina  },
+  /* complex arrays */
+  {"##add.Complex[]Complex[]", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;[]",
+   (SUBR) cops_init,(SUBR)complexa_plus_complexa},
+  {"##mul.Complex[]Complex[]", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;[]",
+   (SUBR) cops_init,(SUBR)complexa_x_complexa},
+  {"##sub.Complex[]Complex[]", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;[]",
+   (SUBR) cops_init,(SUBR)complexa_sub_complexa},
+  {"##div.Complex[]Complex[]", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;[]",
+   (SUBR) cops_init,(SUBR)complexa_div_complexa},
+  {"##add.Complex[]k", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]k",
+   (SUBR) cops_init,(SUBR)complex_plus_scalar},
+  {"##add.kComplex[]", sizeof(COPS1), 0, ":Complex;[]", "k:Complex;[]",
+   (SUBR) cops_init,(SUBR)complex_plus_scalar},
+  {"##mul.Complex[]k", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]k",
+   (SUBR) cops_init,(SUBR)complex_x_scalar},
+  {"##mul.kComplex[]", sizeof(COPS1), 0, ":Complex;[]", "k:Complex;[]",
+   (SUBR) cops_init,(SUBR)complex_x_scalar},
+  {"##div.Complex[]k", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]k",
+   (SUBR) cops_init,(SUBR)complex_div_scalar},   
+  {"##sub.Complex[]k", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]k",
+   (SUBR) cops_init,(SUBR)complex_minus_scalar},
+  {"##sub.kComplex[]", sizeof(COPS1), 0, ":Complex;[]", "k:Complex;[]",
+   (SUBR) cops_init,(SUBR)scalar_minus_complex},
+  {"##add.Complex[]Complex", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;",
+   (SUBR) cops_init,(SUBR)complex_plus_complex},
+  {"##add.ComplexComplex[]", sizeof(COPS1), 0, ":Complex;[]", ":Complex;:Complex;[]",
+   (SUBR) cops_init,(SUBR)complex_plus_complex},
+  {"##mul.Complex[]Complex", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;",
+   (SUBR) cops_init,(SUBR)complex_x_complex},
+  {"##mul.ComplexComplex[]", sizeof(COPS1), 0, ":Complex;[]", ":Complex;:Complex;[]",
+   (SUBR) cops_init,(SUBR)complex_x_complex},
+  {"##div.Complex[]Complex", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;",
+   (SUBR) cops_init,(SUBR)complex_div_complex},   
+  {"##sub.Complex[]Complex", sizeof(COPS1), 0, ":Complex;[]", ":Complex;[]:Complex;",
+   (SUBR) cops_init,(SUBR)complexa_minus_complex},
+  {"##sub.ComplexComplex[]", sizeof(COPS1), 0, ":Complex;[]", ":Complex;:Complex;[]",
+   (SUBR) cops_init,(SUBR)complex_minus_complexa},  
+    
+
+
+  
   { "conj", S(CXOP2R), 0, "k", ":Complex;", NULL,  (SUBR) complex_conj },
   { "conj", S(CXOP2R), 0, "i", ":Complex;", (SUBR) complex_conj },
   { "real", S(CXOP2R), 0, "k", ":Complex;", NULL, (SUBR) complex_real},
