+++ OOps/complex_ops.c
@@ -397,3 +397,425 @@ int32_t real_div_complex(CSOUND *csound, AOP *p) {
 }
 
 
+/** Complex array operators
+ */
+
+/** complex array op real scalar 
+ */
+
+int32_t cops_init(CSOUND *csound, COPS1 *p) {
+  int32_t size;
+  if(IS_ARRAY_ARG(p->a) &&
+     IS_ARRAY_ARG(p->b)){
+    ARRAYDAT *aa = (ARRAYDAT *) p->a; 
+    ARRAYDAT *ab = (ARRAYDAT *) p->b;
+    if(aa->dimensions == 0) return csound->InitError(csound, "array1 unitialised\n");
+    if(ab->dimensions == 0) return csound->InitError(csound, "array2 unitialised\n");
+    size = aa->sizes[0] < ab->sizes[0] ? aa->sizes[0] : ab->sizes[0]; 
+  } else if (IS_ARRAY_ARG(p->a)) {
+    if(((ARRAYDAT *)p->a)->sizes == NULL) return csound->InitError(csound, "array unitialised\n");
+    size = ((ARRAYDAT *)p->a)->sizes[0];
+  }
+  else if (IS_ARRAY_ARG(p->b)) {
+    if(((ARRAYDAT *)p->b)->sizes == NULL) return csound->InitError(csound, "array unitialised\n");
+    size = ((ARRAYDAT *)p->b)->sizes[0];
+  }
+  else size = 1;
+  tabinit(csound, p->out, size, p->h.insdshead);
+  return OK;
+}  
+
+static inline void
+cmplx_sc_prod(COMPLEXDAT *out, COMPLEXDAT *in, MYFLT num, int32_t n) {
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real*num;
+      out[i].imag = in[i].imag*num;
+    } else {
+      out[i].real = in[i].real*num;
+      out[i].imag = in[i].imag;
+    }
+  }
+}
+
+int32_t complex_x_scalar(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array;
+  MYFLT num;
+  if(IS_ARRAY_ARG(p->b)) {
+   array = (ARRAYDAT *) p->b;
+   num = *p->a;
+  } else {
+   array = (ARRAYDAT *) p->a;
+   num = *p->b;
+  } 
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_sc_prod(out,in,num,len);
+  return OK;
+}
+
+static inline void
+cmplx_sc_div(COMPLEXDAT *out, COMPLEXDAT *in, MYFLT num, int32_t n) {
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real/num;
+      out[i].imag = in[i].imag/num;
+    } else {
+      out[i].real = in[i].real/num;
+      out[i].imag = in[i].imag;
+    }
+  }
+}
+
+int32_t complex_div_scalar(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array = (ARRAYDAT *) p->a;
+  MYFLT num = *p->b;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_sc_div(out,in,num,len);
+  return OK;
+}
+
+static inline void
+cmplx_sc_add(COMPLEXDAT *out, COMPLEXDAT *in, MYFLT num, int32_t n) {
+  for(int i = 0; i < n; i++) {
+    if(!in[i].isPolar) {
+      out[i].isPolar = in[i].isPolar;
+      out[i].real = in[i].real + num;
+      out[i].imag = in[i].imag;
+    } else {
+      MYFLT re, im;
+      re = cos(in[i].imag)*in[i].real;  
+      im = sin(in[i].imag)*in[i].real;
+      re += num;
+      out[i].real = HYPOT(re,im);
+      out[i].imag = ATAN2(im,re);
+    }
+  }
+}
+
+int32_t complex_plus_scalar(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array;
+  MYFLT num;
+  if(IS_ARRAY_ARG(p->b)) {
+   array = (ARRAYDAT *) p->b;
+   num = *p->a;
+  } else {
+   array = (ARRAYDAT *) p->a;
+   num = *p->b;
+  } 
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_sc_add(out,in,num,len);
+  return OK;
+}
+
+int32_t scalar_minus_complex(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array = (ARRAYDAT *) p->b;
+  MYFLT num = *p->a;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  for(int i = 0; i < len; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = num - in[i].real;
+      out[i].imag = num;
+    } else {
+      MYFLT re, im;
+      re = cos(in[i].imag)*in[i].real;  
+      im = sin(in[i].imag)*in[i].real;
+      re = num - re;
+      out[i].real = HYPOT(re,im);
+      out[i].imag = ATAN2(im,re);
+    }
+  }
+  return OK;
+}
+
+int32_t complex_minus_scalar(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array = (ARRAYDAT *) p->a;
+  MYFLT num = *p->b;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  for(int i = 0; i < len; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real - num;
+      out[i].imag = in[i].imag;
+    } else {
+      MYFLT re, im;
+      re = cos(in[i].imag)*in[i].real;  
+      im = sin(in[i].imag)*in[i].real;
+      re -= num;
+      out[i].real = HYPOT(re,im);
+      out[i].imag = ATAN2(im,re);
+    }
+  }
+  return OK;
+}
+ 
+
+static inline void
+cmplx_cmplx_prod(COMPLEXDAT *out, COMPLEXDAT *in, COMPLEXDAT num, int32_t n) {
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real*num.real - in[i].imag*num.imag;
+      out[i].imag = in[i].imag*num.real + in[i].real*num.imag;
+    } else {
+      out[i].real = in[i].real*num.real;
+      out[i].imag = in[i].imag + num.imag;
+    }
+  }
+}
+
+int32_t complex_x_complex(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array;
+  COMPLEXDAT num;
+  if(IS_ARRAY_ARG(p->b)) {
+   array = (ARRAYDAT *) p->b;
+   num = *(COMPLEXDAT *)p->a;
+  } else {
+   array = (ARRAYDAT *) p->a;
+   num = *(COMPLEXDAT *)p->b;
+  } 
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_cmplx_prod(out,in,num,len);
+  return OK;
+}
+
+static inline void
+cmplx_cmplx_div(COMPLEXDAT *out, COMPLEXDAT *in, COMPLEXDAT num, int32_t n) {
+  MYFLT d;
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      d = num.real*num.real + num.imag*num.imag;
+      out[i].real = (in[i].real*num.real + in[i].imag*num.imag)/d;
+      out[i].imag = (in[i].imag*num.real - in[i].real*num.imag)/d;
+    } else {
+      out[i].real = in[i].real/num.real;
+      out[i].imag = in[i].imag - num.imag;
+    }
+  }
+}
+
+int32_t complex_div_complex(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array = (ARRAYDAT *) p->a;
+  COMPLEXDAT num = *(COMPLEXDAT *)p->b;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_cmplx_div(out,in,num,len);
+  return OK;
+}
+
+static inline void
+cmplx_cmplx_add(COMPLEXDAT *out, COMPLEXDAT *in, COMPLEXDAT num, int32_t n) {
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real + num.real;
+      out[i].imag = in[i].imag + num.imag;
+    } else {
+      MYFLT re, im;
+      re = cos(in[i].imag)*in[i].real;  
+      im = sin(in[i].imag)*in[i].real;
+      re += num.real;
+      im += num.imag;
+      out[i].real = HYPOT(re,im);
+      out[i].imag = ATAN2(im,re);
+    }
+  }
+}
+
+int32_t complex_plus_complex(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array;
+  COMPLEXDAT num;
+  if(IS_ARRAY_ARG(p->b)) {
+   array = (ARRAYDAT *) p->b;
+   num = *(COMPLEXDAT *)p->a;
+  } else {
+   array = (ARRAYDAT *) p->a;
+   num = *(COMPLEXDAT *)p->b;
+  } 
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_cmplx_add(out,in,num,len);
+  return OK;
+}
+
+int32_t complex_minus_complexa(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array = (ARRAYDAT *) p->b;
+  COMPLEXDAT num = *(COMPLEXDAT *)p->a;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  for(int i = 0; i < len; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = num.real - in[i].real;
+      out[i].imag = num.imag - in[i].imag;
+    } else {
+      MYFLT re, im;
+      re = cos(in[i].imag)*in[i].real;  
+      im = sin(in[i].imag)*in[i].real;
+      re = num.real - re;
+      im = num.imag - im;
+      out[i].real = HYPOT(re,im);
+      out[i].imag = ATAN2(im,re);    
+    }
+  }
+  return OK;
+}
+
+int32_t complexa_minus_complex(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array = (ARRAYDAT *) p->a;
+  COMPLEXDAT num = *(COMPLEXDAT *)p->b;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in = (COMPLEXDAT *) array->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  for(int i = 0; i < len; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real - num.real;
+      out[i].imag = in[i].imag - num.imag;
+    } else {
+      MYFLT re, im;
+      re = cos(in[i].imag)*in[i].real;  
+      im = sin(in[i].imag)*in[i].real;
+      re -= num.real;
+      im -= num.imag;
+      out[i].real = HYPOT(re,im);
+      out[i].imag = ATAN2(im,re);
+    }
+  }
+  return OK;
+}
+
+static inline void
+cmplx_cmplx_proda(COMPLEXDAT *out, COMPLEXDAT *in, COMPLEXDAT *num, int32_t n) {
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real*num[i].real - in[i].imag*num[i].imag;
+      out[i].imag = in[i].imag*num[i].real + in[i].real*num[i].imag;
+    } else {
+      out[i].real = in[i].real*num[i].real;
+      out[i].imag = in[i].imag + num[i].imag;
+    }
+  }
+}
+
+int32_t complexa_x_complexa(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array1, *array2;
+  array2 = (ARRAYDAT *) p->b;
+  array1 = (ARRAYDAT *) p->a;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in1 = (COMPLEXDAT *) array1->data;
+  COMPLEXDAT *in2 = (COMPLEXDAT *) array2->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_cmplx_proda(out,in1,in2,len);
+  return OK;
+}
+
+static inline void
+cmplx_cmplx_diva(COMPLEXDAT *out, COMPLEXDAT *in, COMPLEXDAT* num, int32_t n) {
+  MYFLT d;
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      d = num[i].real*num[i].real + num[i].imag*num[i].imag;
+      out[i].real = (in[i].real*num[i].real + in[i].imag*num[i].imag)/d;
+      out[i].imag = (in[i].imag*num[i].real - in[i].real*num[i].imag)/d;
+    } else {
+      out[i].real = in[i].real/num[i].real;
+      out[i].imag = in[i].imag - num[i].imag;
+    }
+  }
+}
+
+
+int32_t complexa_div_complexa(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array1, *array2;
+  array2 = (ARRAYDAT *) p->b;
+  array1 = (ARRAYDAT *) p->a;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in1 = (COMPLEXDAT *) array1->data;
+  COMPLEXDAT *in2 = (COMPLEXDAT *) array2->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_cmplx_diva(out,in1,in2,len);
+  return OK;
+}
+
+static inline void
+cmplx_cmplx_adda(COMPLEXDAT *out, COMPLEXDAT *in, COMPLEXDAT* num, int32_t n) {
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real + num[i].real;
+      out[i].imag = in[i].imag + num[i].imag;
+    } else {
+      MYFLT re, im;
+      re = cos(in[i].imag)*in[i].real;  
+      im = sin(in[i].imag)*in[i].real;
+      re += num[i].real;
+      im += num[i].imag;
+      out[i].real = HYPOT(re,im);
+      out[i].imag = ATAN2(im,re);
+    }
+  }
+}
+
+int32_t complexa_plus_complexa(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array1, *array2;
+  array2 = (ARRAYDAT *) p->b;
+  array1 = (ARRAYDAT *) p->a;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in1 = (COMPLEXDAT *) array1->data;
+  COMPLEXDAT *in2 = (COMPLEXDAT *) array2->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_cmplx_adda(out,in1,in2,len);
+  return OK;
+}
+
+static inline void
+cmplx_cmplx_suba(COMPLEXDAT *out, COMPLEXDAT *in, COMPLEXDAT* num, int32_t n) {
+  for(int i = 0; i < n; i++) {
+    out[i].isPolar = in[i].isPolar;
+    if(!in[i].isPolar) {
+      out[i].real = in[i].real - num[i].real;
+      out[i].imag = in[i].imag - num[i].imag;
+    } else {
+      MYFLT re, im;
+      re = cos(in[i].imag)*in[i].real;  
+      im = sin(in[i].imag)*in[i].real;
+      re -= num[i].real;
+      im -= num[i].imag;
+      out[i].real = HYPOT(re,im);
+      out[i].imag = ATAN2(im,re);
+    }
+  }
+}
+
+int32_t complexa_sub_complexa(CSOUND *csound, COPS1 *p) {
+  ARRAYDAT *array1, *array2;
+  array2 = (ARRAYDAT *) p->b;
+  array1 = (ARRAYDAT *) p->a;
+  int32_t len = p->out->sizes[0];
+  COMPLEXDAT *in1 = (COMPLEXDAT *) array1->data;
+  COMPLEXDAT *in2 = (COMPLEXDAT *) array2->data;
+  COMPLEXDAT *out = (COMPLEXDAT *) p->out->data;
+  cmplx_cmplx_suba(out,in1,in2,len);
+  return OK;
+}
